requirements.txt
requests==2.31.0
pyyaml==6.0.1
pandas==2.2.2
tenacity==8.2.3
numpy==1.26.4
matplotlib==3.8.4
yfinance==0.2.40